Add loader to run generated wasm code
