+++ static/loader.js
@@ -1,54 +1,13 @@
-
 let wasm;
 
-function init(module) {
-    if (typeof module === 'undefined') {
-        console.error("module is undefined");
-        return;
-    }
-    let result;
-    const imports = {};
-
-    if ((typeof URL === 'function' && module instanceof URL) || typeof module === 'string' || (typeof Request === 'function' && module instanceof Request)) {
-
-        const response = fetch(module);
-        if (typeof WebAssembly.instantiateStreaming === 'function') {
-            result = WebAssembly.instantiateStreaming(response, imports)
-            .catch(e => {
-                return response
-                .then(r => {
-                    if (r.headers.get('Content-Type') != 'application/wasm') {
-                        console.warn("`WebAssembly.instantiateStreaming` failed because your server does not serve wasm with `application/wasm` MIME type. Falling back to `WebAssembly.instantiate` which is slower. Original error:\n", e);
-                        return r.arrayBuffer();
-                    } else {
-                        throw e;
-                    }
-                })
-                .then(bytes => WebAssembly.instantiate(bytes, imports));
-            });
-        } else {
-            result = response
-            .then(r => r.arrayBuffer())
-            .then(bytes => WebAssembly.instantiate(bytes, imports));
-        }
-    } else {
-
-        result = WebAssembly.instantiate(module, imports)
-        .then(result => {
-            if (result instanceof WebAssembly.Instance) {
-                return { instance: result, module };
-            } else {
-                return result;
-            }
-        });
-    }
-    return result.then(({instance, module}) => {
+function init(module_bytecode, imports) {
+    if (!imports) imports = {};
+    return WebAssembly.instantiate(module_bytecode, imports)
+    .then(({instance, module}) => {
         wasm = instance.exports;
-        init.__wbindgen_wasm_module = module;
-
+        //init.__wbindgen_wasm_module = module;
         return wasm;
     });
 }
 
 export default init;
-
